+++ database/migrations/2025_03_15_071117_create_members_table.php
@@ -16,8 +16,8 @@ return new class extends Migration {
             $table->string('name');
             $table->unsignedInteger('ratio');
             $table->json('bank_info')->nullable();
-            $table->decimal('total_payments', 15, 3)->default(0);
-            $table->decimal('total_expenses', 15, 3)->default(0);
+            $table->integer('total_payments')->default(0);
+            $table->integer('total_expenses')->default(0);
             $table->foreignId('group_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
